# openapi.yaml
openapi: 3.1.0
info:
  title: Brain RH API
  version: 1.0.0
  description: |
    API de matching intelligent CV/Offre d'emploi avec analyse LLM.

    **Fonctionnalités:**
    - Parsing de CVs (PDF/DOCX) avec extraction LLM
    - Enrichissement d'offres avec génération must-have/nice-have
    - Matching avec filtrage, scoring et re-ranking LLM
    - Streaming temps-réel (SSE) pour traitements longs
    - Gestion de projets et entreprises
    - Export CSV/JSON des résultats

  contact:
    name: Brain RH API Support
    email: support@brainrh.com
servers:
  - url: http://localhost:8000/api/v1
    description: Serveur de développement
  - url: https://api.brainrh.com/api/v1
    description: Serveur de production
...
paths:
...
  /cvs/{cv_id}:
    get:
      tags: [CVs]
      summary: Récupérer un CV parsé
      operationId: getCV
      parameters:
        - name: cv_id
          in: path
          required: true
          schema:
            type: string
          description: ID du CV (nom de fichier)
      responses:
        '200':
          description: CV trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [CVs]
      summary: Supprimer un CV
      operationId: deleteCV
      parameters:
        - name: cv_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: CV supprimé
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== OFFRES ====================
  /offres:
    post:
      tags: [Offres]
      summary: Créer une offre manuelle
      description: Créer une offre d'emploi avec critères must-have/nice-have définis manuellement
      operationId: createOffre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OffreInput'
      responses:
        '201':
          description: Offre créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offre'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /offres/{offre_id}:
    get:
      tags: [Offres]
      summary: Récupérer une offre
      operationId: getOffre
      parameters:
        - name: offre_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Offre trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offre'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Offres]
      summary: Modifier une offre
      operationId: updateOffre
      parameters:
        - name: offre_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OffreInput'
      responses:
        '200':
          description: Offre modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offre'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Offres]
      summary: Supprimer une offre
      operationId: deleteOffre
      parameters:
        - name: offre_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Offre supprimée
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== MATCHING ====================
...
  /projects:
    get:
      tags: [Projets]
      summary: Lister tous les projets
      operationId: listProjects
      parameters:
        - name: enterprise_id
          in: query
          schema:
            type: string
          description: Filtrer par entreprise
        - name: status
          in: query
          schema:
            type: string
            enum: [actif, archive]
          description: Filtrer par status
      responses:
        '200':
          description: Liste des projets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

    post:
      tags: [Projets]
      summary: Créer un projet
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Projet créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{project_id}:
    get:
      tags: [Projets]
      summary: Récupérer un projet
      operationId: getProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Projet trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Projets]
      summary: Modifier un projet
      operationId: updateProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '200':
          description: Projet modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Projets]
      summary: Supprimer un projet
      operationId: deleteProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Projet supprimé
        '404':
          $ref: '#/components/responses/NotFound'
...
  /enterprises:
    get:
      tags: [Entreprises]
      summary: Lister toutes les entreprises
      operationId: listEnterprises
      responses:
        '200':
          description: Liste des entreprises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enterprise'

    post:
      tags: [Entreprises]
      summary: Créer une entreprise
      operationId: createEnterprise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseInput'
      responses:
        '201':
          description: Entreprise créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          $ref: '#/components/responses/BadRequest'
  /enterprises/{enterprise_id}:
    get:
      tags: [Entreprises]
      summary: Récupérer une entreprise
      operationId: getEnterprise
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entreprise trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Entreprises]
      summary: Modifier une entreprise
      operationId: updateEnterprise
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseInput'
      responses:
        '200':
          description: Entreprise modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Entreprises]
      summary: Supprimer une entreprise
      operationId: deleteEnterprise
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Entreprise supprimée
        '404':
          $ref: '#/components/responses/NotFound'

# ==================== COMPONENTS ====================
components:
  schemas:
...
    Identite:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        telephone:
          type: string
        adresse:
          type: string
        linkedin:
          type: string
        autres_reseaux:
          type: array
          items:
            type: string
    Experience:
      type: object
      properties:
        poste:
          type: string
        entreprise:
          type: string
        lieu:
          type: string
        date_debut:
          type: string
        date_fin:
          type: string
        duree:
          type: string
        missions:
          type: array
          items:
            type: string
    CV:
      type: object
      required:
        - cv
      properties:
        cv:
          type: string
          description: Nom du fichier CV
        identite:
          $ref: '#/components/schemas/Identite'
        titre:
          type: string
        resume_professionnel:
          type: string
        competences_techniques:
          type: array
          items:
            type: string
        competences_transversales:
          type: array
          items:
            type: string
        langues:
          type: array
          items:
            type: string
        experiences_professionnelles:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        formations:
          type: array
          items:
            type: object
            additionalProperties: true
        certifications:
          type: array
          items:
            type: string
        projets:
          type: array
          items:
            type: string
...
    OffreSection:
      type: object
      properties:
        titre:
          type: string
        resume_professionnel:
          type: string
        competences_techniques:
          type: array
          items:
            type: string
        competences_transversales:
          type: array
          items:
            type: string
        langues:
          type: array
          items:
            type: string
        experiences_professionnelles:
          type: array
          items:
            type: object
            additionalProperties: true
        formations:
          type: array
          items:
            type: object
            additionalProperties: true
        certifications:
          type: array
          items:
            type: string
        projets:
          type: array
          items:
            type: string
    Offre:
      type: object
      required:
        - sections
      properties:
        sections:
          $ref: '#/components/schemas/OffreSection'
        must_have:
          type: array
          items:
            type: string
          description: Critères éliminatoires
        nice_have:
          type: array
          items:
            type: string
          description: Critères souhaitables
    OffreInput:
      type: object
      required:
        - sections
      properties:
        sections:
          $ref: '#/components/schemas/OffreSection'
        must_have:
          type: array
          items:
            type: string
        nice_have:
          type: array
          items:
            type: string

    # ==================== MATCHING ====================
...
    Project:
      type: object
      required:
        - id
        - nom
        - created_at
        - last_modified
      properties:
        id:
          type: string
        nom:
          type: string
        enterprise_id:
          type: string
        created_at:
          type: string
          format: date-time
        last_modified:
          type: string
          format: date-time
        status:
          type: string
          enum: [actif, archive]
          default: actif
    ProjectInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
        enterprise_id:
          type: string
        status:
          type: string
          enum: [actif, archive]

    # ==================== ENTREPRISES ====================
    Enterprise:
      type: object
      required:
        - id
        - nom
        - created_at
      properties:
        id:
          type: string
        nom:
          type: string
        secteur:
          type: string
        created_at:
          type: string
          format: date-time
        projects_count:
          type: integer
    EnterpriseInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
        secteur:
          type: string

    # ==================== EVENTS SSE ====================
...
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Code d'erreur machine-readable
        message:
          type: string
          description: Message d'erreur human-readable
        details:
          type: object
          description: Détails additionnels
        status:
          type: integer
          description: Code HTTP
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: INVALID_REQUEST
            message: "Le champ 'files' est requis"
            status: 400
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: NOT_FOUND
            message: "Ressource non trouvée"
            status: 404
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API pour authentification (à implémenter si nécessaire)

# Note: Pas de sécurité activée pour le moment (développement local)
# security:
#   - ApiKeyAuth: []